mysliderproject/requirements.txt
Django==1.8
django-bootstrap3==7.0.1
django-filer==1.2.4
django-hero-slider==1.3
django-hvad==1.5.1
django-libs==1.67.5
django-mptt==0.8.5
django-polymorphic==0.8.1
easy-thumbnails==2.3
Pillow==3.3.0
Unidecode==0.4.19
